refactor: replace per-aggregate repos with generic Repository<T, TId>
- Add Abstractions layer (IDatabaseConnection, IDatabaseUnitOfWork, IEntityMapping)
- Single generic Repository<T, TId> class for all aggregates
- Dapper fully behind IDatabaseConnection interface
- 5 hand-written repositories replaced by 5 IEntityMapping configs
- ToolRepository subclass handles ScarcityByLocation child table
- DapperUnitOfWork auto-initializes connection+transaction on first access
- No explicit type casts in mappings (typed DTOs via type handlers)
