+++ src/TSQR.ToolLibrary.Infrastructure/Abstractions/IDatabaseUnitOfWork.cs
@@ -1,0 +1,8 @@
+using TSQR.ToolLibrary.Domain;
+
+namespace TSQR.ToolLibrary.Infrastructure.Abstractions;
+
+public interface IDatabaseUnitOfWork : IUnitOfWork
+{
+    IDatabaseConnection Connection { get; }
+}
