+++ src/TSQR.ToolLibrary.Infrastructure/Abstractions/Repository.cs
@@ -1,0 +1,46 @@
+using TSQR.ToolLibrary.Domain;
+
+namespace TSQR.ToolLibrary.Infrastructure.Abstractions;
+
+public class Repository<TEntity, TId> : IRepository<TEntity, TId>
+    where TEntity : Entity<TId>, IAggregateRoot
+    where TId : ValueObject
+{
+    private readonly IDatabaseUnitOfWork _uow;
+    private readonly IEntityMapping<TEntity> _mapping;
+
+    public Repository(IDatabaseUnitOfWork uow, IEntityMapping<TEntity> mapping)
+    {
+        _uow = uow;
+        _mapping = mapping;
+    }
+
+    public IUnitOfWork UnitOfWork => _uow;
+    protected IDatabaseConnection Database => _uow.Connection;
+
+    public virtual async Task<TEntity?> GetByIdAsync(TId id, CancellationToken cancellationToken = default)
+    {
+        return await _mapping.GetByIdAsync(_uow.Connection, id);
+    }
+
+    public virtual async Task AddAsync(TEntity entity, CancellationToken cancellationToken = default)
+    {
+        var db = _uow.Connection;
+        var id = await db.ExecuteScalarAsync<int>(
+            _mapping.InsertSql, _mapping.ToInsertParameters(entity));
+
+        entity.SetAssignedId((TId)Activator.CreateInstance(typeof(TId), [id])!);
+    }
+
+    public virtual void Update(TEntity entity)
+    {
+        var db = _uow.Connection;
+        db.Execute(_mapping.UpdateSql, _mapping.ToUpdateParameters(entity));
+    }
+
+    public virtual void Delete(TEntity entity)
+    {
+        var db = _uow.Connection;
+        db.Execute(_mapping.DeleteSql, new { Id = entity.Id });
+    }
+}
