+++ src/TSQR.ToolLibrary.WebApi/Program.cs
@@ -5,7 +5,9 @@ using TSQR.ToolLibrary.Domain.Aggregates.InventoryAggregate;
 using TSQR.ToolLibrary.Domain.Aggregates.ReservationAggregate;
 using TSQR.ToolLibrary.Domain.Aggregates.MaintenanceAggregate;
 using TSQR.ToolLibrary.Application.Tool.Commands;
+using TSQR.ToolLibrary.Infrastructure.Abstractions;
 using TSQR.ToolLibrary.Infrastructure.Dapper;
+using TSQR.ToolLibrary.Infrastructure.Dapper.Mappings;
 using TSQR.ToolLibrary.Infrastructure.Dapper.Repositories;
 
 TypeHandlerRegistrations.EnsureRegistered();
@@ -17,16 +19,22 @@ builder.Services.AddOpenApi();
 var connectionString = builder.Configuration.GetConnectionString("DefaultConnection")
     ?? throw new InvalidOperationException("Connection string 'DefaultConnection' not found.");
 
-builder.Services.AddScoped(_ => new DapperUnitOfWork(connectionString));
+builder.Services.AddScoped<IDatabaseUnitOfWork>(_ => new DapperUnitOfWork(connectionString));
 
 builder.Services.AddMediatR(cfg =>
     cfg.RegisterServicesFromAssembly(typeof(RegisterToolCommand).Assembly));
 
+builder.Services.AddSingleton<IEntityMapping<InventoryItem>, InventoryItemMapping>();
+builder.Services.AddSingleton<IEntityMapping<Member>, MemberMapping>();
+builder.Services.AddSingleton<IEntityMapping<Reservation>, ReservationMapping>();
+builder.Services.AddSingleton<IEntityMapping<MaintenanceRecord>, MaintenanceRecordMapping>();
+builder.Services.AddSingleton<IEntityMapping<Tool>, ToolMapping>();
+
+builder.Services.AddScoped<IRepository<InventoryItem, InventoryItemId>, Repository<InventoryItem, InventoryItemId>>();
+builder.Services.AddScoped<IRepository<Member, MemberId>, Repository<Member, MemberId>>();
+builder.Services.AddScoped<IRepository<Reservation, ReservationId>, Repository<Reservation, ReservationId>>();
+builder.Services.AddScoped<IRepository<MaintenanceRecord, MaintenanceRecordId>, Repository<MaintenanceRecord, MaintenanceRecordId>>();
 builder.Services.AddScoped<IRepository<Tool, ToolId>, ToolRepository>();
-builder.Services.AddScoped<IRepository<InventoryItem, InventoryItemId>, InventoryItemRepository>();
-builder.Services.AddScoped<IRepository<Member, MemberId>, MemberRepository>();
-builder.Services.AddScoped<IRepository<Reservation, ReservationId>, ReservationRepository>();
-builder.Services.AddScoped<IRepository<MaintenanceRecord, MaintenanceRecordId>, MaintenanceRecordRepository>();
 
 var app = builder.Build();
